Add: settings in work v.2

+++ src/topseudo/config/settings.py
@@ -1,17 +1,34 @@
 import json
+from os.path import exists
 
 config_structure = {
-    "API-key": "",
+    "key": "",
     "localisation": "",
     "pseudocode-type": ""
 }
 
 
-def create_config_file():
-    with open("settings.json", "w") as file:
-        pass
+class Settings:
+    __instance = None
+    def __new__(cls, *args, **kwargs):
+        if cls.__instance is None:
+            cls.__instance = super(Settings, cls).__new__(cls)
+        return cls.__instance
+
+    def __setitem__(self, key, value):
+        with open("settings.json", "w+") as file:
+            settings = json.load(file)
+            settings[key] = value
+
+    @staticmethod
+    def init_settings():
+        if exists("settings.json"):
+            raise FileExistsError
+        else:
+            with open("settings.json", "w") as file:
+                json.dump(config_structure, file, indent=4)
 
 
 
 
-create_config_file()
+Settings.init_settings()
